+++ sim_api/simulate.jl
@@ -1,50 +1,12 @@
 using JSON
-using Plots
+using Logging
+
+include("./resie/src/resie_logger.jl")
+using .Resie_Logger
+
+using Resie
 
 include("util.jl")
-
-function julia_set(sx, sy, nx, ny, c=-0.744 + 0.148im)::Array{UInt8, 2}
-	iterations_till_divergence = Array{UInt8, 2}(undef, nx, ny)
-    step_x = sx / nx
-    step_y = sy / ny
-    for ix in range(1, length=nx)
-		for iy in range(1, length=ny)
-			iterations_till_divergence[ix,iy] = trunc(UInt8, 0)
-			z = ix * step_x - sx * 0.5 + (iy * step_y - sy * 0.5)*1im
-			for i in 0:254
-				z = z^2 + c
-				if abs(z) > 4.0
-					iterations_till_divergence[ix,iy] = trunc(UInt8, i)
-					break
-                end
-            end
-        end
-    end
-
-	return iterations_till_divergence
-end
-
-function plot_julia_set(sx, sy, values::Array{UInt8, 2}, working_dir::String)
-    nx, ny = size(values)
-    x_min, x_max = -sx/2, sx/2
-    y_min, y_max = -sy/2, sy/2
-    heatmap(
-        range(x_min, x_max; length=nx),
-        range(y_min, y_max; length=ny),
-        values',
-        color=:viridis,
-        aspect_ratio=:equal,
-        title="Julia Set",
-        xlabel="Re",
-        ylabel="Im",
-        dpi=1000
-    )
-    alias = update_file_index(working_dir, "julia_set.png")
-    png(joinpath(working_dir, alias))
-    # rename saved file without file ending as the png function always adds the file ending
-    # but the file index does not expect it
-    mv(joinpath(working_dir, alias * ".png"), joinpath(working_dir, alias))
-end
 
 function simulate(working_dir)
     config_file = joinpath(working_dir, "aliased_config.json")
@@ -53,14 +15,35 @@ function simulate(working_dir)
         return
     end
 
-    try
-        config = JSON.parsefile(config_file)
-        c = Float64(config["c_re"]) + Float64(config["c_im"]) * 1im
-        plot_julia_set(3.0, 2.0, julia_set(3.0, 2.0, 5000, 5000, c), working_dir)
-    catch e
-        println("Error: Could not parse JSON in config file: $e")
-        return
-    end
+    log_to_console = false
+    log_to_file = true
+    general_logfile_path = joinpath(working_dir, "./logfile_general.log")
+    balanceWarn_logfile_path = joinpath(working_dir, "./logfile_balanceWarn.log")
+    min_log_level = Resie_Logger.Logging.Info
+    logger = Resie_Logger.start_logger(log_to_console,
+                                       log_to_file,
+                                       general_logfile_path,
+                                       balanceWarn_logfile_path,
+                                       min_log_level,
+                                       config_file)
 
-    println("Success: Simulation complete")
+    run_ID = uuid1() # should we rather use the sim API run ID?
+    try
+        with_logger(logger) do
+            Resie.load_and_run(config_file, run_ID)
+        end
+    catch exc
+        io = IOBuffer()
+        showerror(io, exc)
+        print(io, stacktrace(catch_backtrace()))
+        msg = String(take!(io))
+        @error msg
+    finally
+        Resie.close_run(run_ID)
+        with_logger(logger) do
+            # the close_logger makes logs itself, which is why we need to use with_logger
+            # or else these messages pop up in the global logger
+            Resie_Logger.close_logger(logger)
+        end
+    end
 end