+++ webapp/simon_webapp/app.py
@@ -22,6 +22,10 @@ if os.environ.get("FLASK_ENV") == "development":
 APP_ROOT = Path(__file__).resolve().parent.parent
 APP_CONFIG_PATH = APP_ROOT / "webapp_config.yml"
 
+# cached value for the ReSiE version. It will be filled on first request defaulting to None
+# indicating we haven't read it yet.
+RESIE_VERSION: str | None = None
+
 # ---------------------------------------------------------------------------
 # App construction
 # ---------------------------------------------------------------------------
@@ -60,7 +64,21 @@ def index():
     if "user_id" not in session:
         session["user_id"] = "__anonymous__"
         session["nextcloud_authorized"] = False
-    return render_template("index.html", session=session), 200
+
+    # lazily load and cache the version string
+    global RESIE_VERSION
+    if RESIE_VERSION is None:
+        response = requests.get(
+            app.config["sim_api"]["endpoint"] + "resie_version",
+            timeout=2, # short timeout because we expect a speedy response and don't want to
+                       # keep the user waiting long
+        )
+        if response.status_code == 200:
+            RESIE_VERSION = response.json()["version"]
+
+    api_root = app.config["SIMON_API_ROOT"]
+    return render_template("index.html", session=session,
+                           api_root=api_root, resie_version=RESIE_VERSION), 200
 
 @app.route("/nextcloud_login", methods=["GET"])
 def nextcloud_login():
@@ -274,34 +292,38 @@ def fetch_results(run_id):
     Request arguments (route):
         - run_id -> str: The ID of the run for which to fetch results
     Request arguments (json):
+        - filename -> str: The name of the file to fetch
         - destination_dir -> str: The directory path on NC to where the results should
             be uploaded. This should be a NC-encoded path.
 
     Response (ByteStream): The results file
     """
-    file_name = "julia_set.png" # will be dynamic later
+    if "filename" not in request.json:
+        return jsonify({"error": "No file name specified"}), 400
     if "destination_dir" not in request.json:
         return jsonify({"error": "No destination specified"}), 400
 
     # fetch file from sim API
     sim_response = requests.post(
         app.config["sim_api"]["endpoint"] + "download_file/" + run_id,
-        json={"filename": file_name},
+        json={"filename": request.json["filename"]},
         timeout=app.config["sim_api"]["timeout"],
         headers={"Authorization": "Bearer " + app.config["sim_api"]["api_key"]}
     )
     if not sim_response.ok:
-        return jsonify({"error": "Could not fetch results from sim API"}), 500
+        msg = f"Could not fetch result file {request.json["filename"]} from sim API"
+        return jsonify({"error": msg}), 500
 
     # upload to NC
     user = quote(session["user_id"])
-    destination = request.json["destination_dir"] + "/" + file_name
+    destination = request.json["destination_dir"] + "/" + request.json["filename"]
     url = app.config["NEXTCLOUD_API_BASE_URL"] + "remote.php/dav/files/" + user \
         + "/" + encode_nc_path(destination)
     nc_response = ensure_request(url, app, method="PUT", data=sim_response.content)
 
     if not nc_response.ok:
-        return jsonify({"error": "Could not upload results to NextCloud"}), 500
+        msg = f"Could not upload result file {request.json["filename"]} to NextCloud"
+        return jsonify({"error": msg}), 500
 
     # return results so the frontend can display them too
     return sim_response.content, 200